Move viewport function to editor utility
