+++ src/utils/editor/index.js
@@ -158,6 +158,35 @@ function isVisible(codeMirror: any, top: number, left: number) {
   return inXView && inYView;
 }
 
+export function getLocationsInViewport(_editor: any) {
+  // Get scroll position
+  const charWidth = _editor.codeMirror.defaultCharWidth();
+  const scrollArea = _editor.codeMirror.getScrollInfo();
+  const { scrollLeft } = _editor.codeMirror.doc;
+  const rect = _editor.codeMirror.getWrapperElement().getBoundingClientRect();
+  const topVisibleLine = _editor.codeMirror.lineAtHeight(rect.top, "window");
+  const bottomVisibleLine = _editor.codeMirror.lineAtHeight(
+    rect.bottom,
+    "window"
+  );
+
+  // Update state which should trigger appropriate re-renders
+  const leftColumn = Math.floor(scrollLeft > 0 ? scrollLeft / charWidth : 0);
+  const rightPosition = scrollLeft + (scrollArea.clientWidth - 30);
+  const rightCharacter = Math.floor(rightPosition / charWidth);
+
+  return {
+    start: {
+      line: topVisibleLine,
+      column: leftColumn
+    },
+    end: {
+      line: bottomVisibleLine,
+      column: rightCharacter
+    }
+  };
+}
+
 export function markText(_editor: any, className, { start, end }: EditorRange) {
   return _editor.codeMirror.markText(
     { ch: start.column, line: start.line },
